Replace Verification Token with Signing Secret
The Verification Token is not required when configuring Slack in Azure. In fact, you can see in the screenshot that immediately follows these instructions that the Azure UI doesn't ask for it. Instead, it asks for the Signing Secret.

![](https://docs.microsoft.com/en-us/azure/bot-service/media/channels/slack-submitcredentials.png?view=azure-bot-service-4.0)

This makes sense because the verification token is deprecated according to Slack.

> This deprecated Verification Token can still be used to verify that requests come from Slack, but we strongly recommend using the above, more secure, signing secret instead.